+++ include/ippp/core/sampling/BridgeSampling.hpp
@@ -33,7 +33,7 @@ class BridgeSampling : public Sampling<dim> {
   public:
     BridgeSampling(const std::shared_ptr<Environment> &environment, const std::shared_ptr<CollisionDetection<dim>> &collision,
                    const std::shared_ptr<TrajectoryPlanner<dim>> &trajectory, const std::shared_ptr<Sampler<dim>> &sampler,
-                   const unsigned int attempts = 10, const double distance = 15);
+                   const size_t attempts = 10, const double distance = 15);
 
     Vector<dim> getSample() override;
 
@@ -59,7 +59,7 @@ template <unsigned int dim>
 BridgeSampling<dim>::BridgeSampling(const std::shared_ptr<Environment> &environment,
                                     const std::shared_ptr<CollisionDetection<dim>> &collision,
                                     const std::shared_ptr<TrajectoryPlanner<dim>> &trajectory,
-                                    const std::shared_ptr<Sampler<dim>> &sampler, const unsigned int attempts,
+                                    const std::shared_ptr<Sampler<dim>> &sampler, const size_t attempts,
                                     const double distance)
     : Sampling<dim>("BridgeSampling", environment, collision, trajectory, sampler, attempts), m_distance(distance) {
 }
@@ -72,7 +72,7 @@ BridgeSampling<dim>::BridgeSampling(const std::shared_ptr<Environment> &environm
 */
 template <unsigned int dim>
 Vector<dim> BridgeSampling<dim>::getSample() {
-    unsigned int count = 0;
+    size_t count = 0;
     Vector<dim> sample1, sample2, ray;
     do {
         sample1 = m_sampler->getSample();
