+++ include/ippp/core/sampling/GaussianDistSampling.hpp
@@ -33,7 +33,7 @@ class GaussianDistSampling : public Sampling<dim> {
   public:
     GaussianDistSampling(const std::shared_ptr<Environment> &environment, const std::shared_ptr<CollisionDetection<dim>> &collision,
                      const std::shared_ptr<TrajectoryPlanner<dim>> &trajectory, const std::shared_ptr<Sampler<dim>> &sampler,
-                     const unsigned int attempts = 10, const double maxDist = 15);
+                     const size_t attempts = 10, const double maxDist = 15);
 
     Vector<dim> getSample() override;
     Vector<dim> getSample(const Vector<dim> &prevSample) override;
@@ -61,7 +61,7 @@ template <unsigned int dim>
 GaussianDistSampling<dim>::GaussianDistSampling(const std::shared_ptr<Environment> &environment,
                                         const std::shared_ptr<CollisionDetection<dim>> &collision,
                                         const std::shared_ptr<TrajectoryPlanner<dim>> &trajectory,
-                                        const std::shared_ptr<Sampler<dim>> &sampler, const unsigned int attempts,
+                                        const std::shared_ptr<Sampler<dim>> &sampler, const size_t attempts,
                                         const double maxDist)
     : Sampling<dim>("GaussianDistSampling", environment, collision, trajectory, sampler, attempts), m_maxDist(maxDist) {
 }
@@ -88,7 +88,7 @@ Vector<dim> GaussianDistSampling<dim>::getSample() {
 template <unsigned int dim>
 Vector<dim> GaussianDistSampling<dim>::getSample(const Vector<dim> &prevSample) {
     Vector<dim> ray, sample;
-    for (unsigned int count = 0; count < m_attempts; ++count) {
+    for (size_t count = 0; count < m_attempts; ++count) {
         ray = m_sampler->getRandomRay();
         ray *= m_maxDist * m_sampler->getRandomNumber();
         sample = prevSample + ray;
